add reverse implementation; update tests
